Final iOS changes, Bulk REST API, fix in deactivated users list api

+++ iOS/SundayFriends/Models/Transaction.swift
@@ -20,7 +20,7 @@ class Transaction: Codable {
   
   var type: Int?
   var amount: Double?
-  var balanceAfterAction: Int?
+  var balanceAfterAction: Double?
   var time: String?
   
   init() {}
